fix: TripPlanView Type Mismatch
